bugfix: fix group recommend request
